--- Software/GuitarPedal/Effect-Modules/Nam/NamA2JCM2000Daisy48.h
@@ -1,2629 +1,0 @@
-#pragma once
-/*
-    NamA2JCM2000Daisy48_minram_v16_k6_current_direct.h
-
-    Single-header, allocation-free A2 Nano WaveNet runtime for Daisy / STM32H7.
-
-    Model embedded:
-      Marshall JCM2000 DSL100_A2.nam
-      SlimmableContainer low-CPU submodel: A2 WaveNet, 3 channels, 1871 weights.
-      Block size: 48 samples. Mono input -> mono output.
-
-    Usage:
-      - Call load_default_model() once from setup / Init, not from the audio callback.
-      - Call process_block_48(in, out) from the audio callback with exactly 48 samples.
-      - For stereo processing, create two A2State objects and one shared A2SharedWeights.
-
-    v16 change:
-      - branchless leaky from v15
-      - K=6 dual kernel uses direct current tap for non-layer-0 layers
-
-    Hot-path design:
-      - no heap
-      - no std::vector
-      - no modulo / division in the audio callback
-      - C=3 hard-unrolled matrix-vector math
-      - 16-tap head uses bitmask wrapping
-      - hybrid ring histories: pow2 rings for small layers, exact rings for large layers
-      - flattened contiguous history arena
-*/
-
-#include <algorithm>
-#include <array>
-#include <cstddef>
-#include <cstdint>
-#include <cstring>
-
-#ifndef NAM_A2_ALIGN32
-#define NAM_A2_ALIGN32 alignas(32)
-#endif
-
-// v8 storage policy:
-//   - NAM_A2_HOT_DATA: packed weights/biases; defaults to DSY_DTCMRAM when available.
-//   - NAM_A2_HOT_STATE_DATA: small hot work buffers; defaults to DSY_DTCMRAM when available.
-//   - NAM_A2_STATE_DATA: large history buffer; defaults to normal SRAM.
-//
-// This keeps DTCM below ~15 KB while moving the high-traffic 48-sample
-// ping-pong/head buffers into DTCM. Large dilation history stays in SRAM.
-#ifndef NAM_A2_HOT_DATA
-#define NAM_A2_HOT_DATA __attribute__((section(".dtcmram_bss")))
-#endif
-
-#ifndef NAM_A2_STATE_DATA
-#define NAM_A2_STATE_DATA
-#endif
-
-#ifndef NAM_A2_HOT_STATE_DATA
-#define NAM_A2_HOT_STATE_DATA __attribute__((section(".dtcmram_bss")))
-#endif
-
-#ifndef NAM_A2_MODEL_DATA
-#define NAM_A2_MODEL_DATA __attribute__((section(".qspiflash_text")))
-#endif
-
-// Minimum-SRAM / BOOT_SRAM build: do not force-inline large DSP helpers.
-// In BOOT_SRAM, .text consumes SRAM, so code size matters more than speed.
-#ifndef NAM_A2_FORCE_INLINE
-#define NAM_A2_FORCE_INLINE inline
-#endif
-
-#ifndef NAM_A2_NOINLINE
-#if defined(__GNUC__) || defined(__clang__)
-#define NAM_A2_NOINLINE __attribute__((noinline))
-#else
-#define NAM_A2_NOINLINE
-#endif
-#endif
-
-#ifndef NAM_A2_TINY_INLINE
-#if defined(__GNUC__) || defined(__clang__)
-#define NAM_A2_TINY_INLINE inline __attribute__((always_inline))
-#else
-#define NAM_A2_TINY_INLINE inline
-#endif
-#endif
-
-#ifndef NAM_A2_ENABLE_PREFETCH
-#define NAM_A2_ENABLE_PREFETCH 0
-#endif
-
-#if NAM_A2_ENABLE_PREFETCH && (defined(__GNUC__) || defined(__clang__))
-  #define NAM_A2_PREFETCH_R(PTR) __builtin_prefetch((PTR), 0, 1)
-#else
-  #define NAM_A2_PREFETCH_R(PTR) do {} while (false)
-#endif
-
-namespace nam_a2_daisy
-{
-static constexpr int kBlockSize = 48;
-static constexpr int kChannels = 3;
-static constexpr int kNumLayers = 23;
-static constexpr int kHeadKernel = 16;
-static constexpr float kLeakySlope = 0.01f;
-
-inline constexpr int kKernelSizes[kNumLayers] = {
-    6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 15, 15, 6, 6, 6, 6, 6, 6, 6
-};
-
-inline constexpr int kDilations[kNumLayers] = {
-    1, 3, 7, 17, 41, 101, 239, 1, 3, 7, 17, 41, 101, 239, 1, 13, 1, 3, 7, 17, 41, 101, 239
-};
-
-static constexpr int kKernelSum = 156;
-static constexpr int kConvWeightCount = kKernelSum * kChannels * kChannels; // 1404
-static constexpr int kLayer1x1WeightCount = kNumLayers * kChannels * kChannels; // 207
-static constexpr int kA2WeightCount = 1871;
-
-// CPU-aggressive compact history policy.
-// Use exact ring sizes plus one safety column. The extra column allows the
-// K=6 dual-sample kernel to write samples n and n+1 before accumulating both
-// without overwriting the oldest tap needed by sample n.
-// Cost versus exact rings: 23 layers * 3 floats = 276 bytes.
-static constexpr int exact_cols_for_layer(int i) noexcept
-{
-    return (kKernelSizes[i] - 1) * kDilations[i] + 2;
-}
-
-static constexpr int compute_exact_history_columns() noexcept
-{
-    int total = 0;
-    for (int i = 0; i < kNumLayers; ++i)
-        total += exact_cols_for_layer(i);
-    return total;
-}
-
-static constexpr int kHistoryColumns = compute_exact_history_columns(); // 6377
-static constexpr int kHistoryFloats  = kHistoryColumns * kChannels;     // 19131
-static constexpr int kHistoryBytes   = kHistoryFloats * 4;              // 76524
-
-NAM_A2_MODEL_DATA NAM_A2_ALIGN32 inline constexpr float kJcm2000A2NanoWeights[kA2WeightCount] = {
-    0.608543336f,
-    -1.31986153f,
-    -2.18866348f,
-    0.0374212898f,
-    -0.00201730197f,
-    -0.066708073f,
-    -0.118407287f,
-    -0.139832318f,
-    -0.112117305f,
-    -0.034495566f,
-    0.0154172657f,
-    0.0897956416f,
-    0.135134816f,
-    0.133655578f,
-    0.120441698f,
-    -0.0268766955f,
-    0.0346290246f,
-    0.112851948f,
-    0.150970906f,
-    0.121884391f,
-    0.131064922f,
-    -0.291776091f,
-    -0.0511414073f,
-    -0.0390029848f,
-    -0.0563128963f,
-    -0.132458478f,
-    0.15485397f,
-    0.0624424033f,
-    -0.0799126476f,
-    -0.178828686f,
-    -0.143318817f,
-    -0.0853131637f,
-    -0.257001668f,
-    0.130787894f,
-    0.0623608716f,
-    -0.0169348102f,
-    0.117944978f,
-    0.201342091f,
-    -0.366169214f,
-    0.0358611755f,
-    -0.0609737188f,
-    -0.485390931f,
-    -0.261264682f,
-    -0.129606068f,
-    -0.55240345f,
-    -0.163268492f,
-    -0.0309630446f,
-    -0.205002487f,
-    -0.430415213f,
-    -0.330861777f,
-    0.34435764f,
-    -0.138795495f,
-    -0.102927446f,
-    0.0395794883f,
-    -0.561444998f,
-    -0.629668117f,
-    0.31126225f,
-    -0.39296779f,
-    -0.17225273f,
-    0.213426068f,
-    -0.115694508f,
-    0.80632925f,
-    0.00313426764f,
-    0.127661467f,
-    -0.245863304f,
-    0.689281881f,
-    -0.0553685687f,
-    0.174693629f,
-    0.35563609f,
-    -0.450506926f,
-    0.803920507f,
-    -0.503506005f,
-    0.121809833f,
-    -0.087087892f,
-    0.0274298303f,
-    -0.336213976f,
-    0.0893574581f,
-    -0.0474531315f,
-    -0.373316199f,
-    0.196813866f,
-    0.427812546f,
-    0.235342354f,
-    0.256163001f,
-    0.656724453f,
-    -0.0832065865f,
-    -0.643639743f,
-    -0.227131233f,
-    0.65528965f,
-    0.293938398f,
-    0.53885597f,
-    0.233894765f,
-    -0.982973814f,
-    -0.687652528f,
-    -0.293847442f,
-    0.120657086f,
-    -0.00945101865f,
-    -0.326710939f,
-    0.149964839f,
-    -0.181934938f,
-    0.0106276022f,
-    -0.14354378f,
-    -0.211257935f,
-    -0.0938126221f,
-    -0.366010845f,
-    -0.0621241108f,
-    0.0598559566f,
-    0.251854807f,
-    0.30567956f,
-    0.219366729f,
-    -0.627269506f,
-    -0.790167868f,
-    -0.183608294f,
-    0.569883645f,
-    -0.171442673f,
-    -0.326408684f,
-    0.165726498f,
-    0.349884719f,
-    -0.124600902f,
-    0.296592295f,
-    0.282962143f,
-    0.123509251f,
-    -0.660983443f,
-    -0.300487369f,
-    0.512605429f,
-    0.296590388f,
-    0.60770911f,
-    0.0432264656f,
-    -0.661405027f,
-    -0.369726837f,
-    0.125543088f,
-    0.154283598f,
-    0.319914013f,
-    0.752645671f,
-    0.86202389f,
-    1.39872825f,
-    -0.311590761f,
-    0.219351932f,
-    -0.398836195f,
-    -0.08176741f,
-    0.466183841f,
-    0.0399382934f,
-    -1.64571488f,
-    -0.802629948f,
-    -1.26564479f,
-    0.401579857f,
-    -0.229691178f,
-    0.325131804f,
-    0.459905237f,
-    -0.0621894076f,
-    0.0692999512f,
-    -0.12987636f,
-    -0.25373888f,
-    0.0388107151f,
-    0.200060904f,
-    0.00244129007f,
-    -0.166579589f,
-    -0.136093393f,
-    -0.414145887f,
-    0.131896302f,
-    -0.0351385102f,
-    -0.0216673966f,
-    0.0430510789f,
-    0.146542504f,
-    -0.016362587f,
-    0.576408863f,
-    0.179651156f,
-    0.153832749f,
-    0.0489986166f,
-    -0.409872323f,
-    -0.139308736f,
-    -0.178587139f,
-    0.0597665794f,
-    0.137351543f,
-    0.0339539945f,
-    -0.0837552696f,
-    -0.15682818f,
-    -0.026038656f,
-    -0.0582104288f,
-    -0.137537017f,
-    -0.0562801249f,
-    0.0950856879f,
-    -0.0170039162f,
-    -0.500931382f,
-    -0.117224514f,
-    0.0546867587f,
-    -0.0705188736f,
-    0.0184265878f,
-    -0.235724717f,
-    -0.0240532607f,
-    0.0982968956f,
-    0.0622842424f,
-    0.118095778f,
-    0.350043893f,
-    0.147265375f,
-    0.541695774f,
-    0.00750179216f,
-    0.0571568981f,
-    -0.0011243812f,
-    -0.0335460603f,
-    0.0396859683f,
-    -0.66317904f,
-    -0.265434176f,
-    -0.243682206f,
-    -0.205001295f,
-    -0.63324368f,
-    0.183750212f,
-    -0.287538528f,
-    -0.272999406f,
-    -0.0735127106f,
-    0.125303328f,
-    -0.0761516318f,
-    -0.288064808f,
-    0.164892033f,
-    -0.728391886f,
-    0.361355186f,
-    0.755949676f,
-    0.167163238f,
-    -0.119689725f,
-    0.484695584f,
-    -0.358009696f,
-    -0.30119133f,
-    -0.298988819f,
-    -0.466643184f,
-    -0.43675667f,
-    -0.234903768f,
-    0.040511433f,
-    0.124131233f,
-    0.236878395f,
-    0.0806881338f,
-    0.158805892f,
-    0.219455436f,
-    0.130671069f,
-    0.14121823f,
-    0.0373473167f,
-    0.195541814f,
-    -0.335086167f,
-    -0.182526469f,
-    -0.209066898f,
-    -0.275267422f,
-    -0.216116905f,
-    -0.35332188f,
-    -0.44403702f,
-    -0.0872882381f,
-    0.108113833f,
-    0.0765649304f,
-    0.178761348f,
-    0.147121295f,
-    0.147895217f,
-    0.194617003f,
-    0.0813245028f,
-    -0.0507537425f,
-    -0.0794906616f,
-    0.215673894f,
-    -0.324079126f,
-    -0.114231877f,
-    0.0902679339f,
-    0.200704113f,
-    -0.0593311675f,
-    -0.010846517f,
-    0.250623226f,
-    0.149037138f,
-    0.026338201f,
-    -0.208835512f,
-    -0.155417427f,
-    -0.0470134802f,
-    -0.107873328f,
-    0.136588454f,
-    0.0955882892f,
-    0.275870293f,
-    0.352009892f,
-    0.475813687f,
-    0.22758171f,
-    -0.688879192f,
-    -0.270160735f,
-    -0.244915262f,
-    0.00433202274f,
-    -0.0610006601f,
-    -0.0482791439f,
-    -0.597907364f,
-    0.137493193f,
-    0.190610066f,
-    -0.619359851f,
-    -0.41156286f,
-    -0.27912122f,
-    -0.00928153377f,
-    -0.53576082f,
-    -0.374888003f,
-    -0.3938438f,
-    -0.299241185f,
-    -0.302194148f,
-    0.485966176f,
-    -0.0165617522f,
-    0.0412888005f,
-    -0.0445762947f,
-    0.0262022577f,
-    -0.44686842f,
-    -0.0597838089f,
-    0.253084719f,
-    0.240518823f,
-    0.048320625f,
-    0.140403077f,
-    0.210288227f,
-    -0.0511373505f,
-    -0.0690610036f,
-    0.0618543737f,
-    -0.0306279045f,
-    -0.171485424f,
-    0.178440303f,
-    0.0192184076f,
-    -0.111802764f,
-    0.0864628479f,
-    -0.173110858f,
-    -0.32402283f,
-    -0.610472381f,
-    -0.0459541082f,
-    -0.204085767f,
-    -0.172386497f,
-    0.0522137322f,
-    -0.0169353373f,
-    -0.157117695f,
-    -0.0112549793f,
-    0.0810053423f,
-    -0.0853869915f,
-    -0.010685144f,
-    0.0792928711f,
-    0.362227947f,
-    -0.482592165f,
-    0.1364398f,
-    -0.105080225f,
-    0.0332491733f,
-    0.0295423139f,
-    -0.662611485f,
-    -0.122434676f,
-    -0.217517406f,
-    -0.0894109011f,
-    -0.0412064046f,
-    0.0352185145f,
-    0.019426465f,
-    0.0157519113f,
-    -0.0995275974f,
-    0.112410851f,
-    0.116334915f,
-    0.108983263f,
-    0.403808981f,
-    -0.398734003f,
-    -0.123356879f,
-    -0.0586870387f,
-    0.277655184f,
-    0.0128429281f,
-    0.221901879f,
-    -0.553149939f,
-    0.134451255f,
-    -0.112612568f,
-    -0.0910788849f,
-    -0.435793042f,
-    0.0426434614f,
-    0.292331606f,
-    -0.441670299f,
-    -0.634085655f,
-    -0.505041361f,
-    -0.151584685f,
-    0.0226566792f,
-    0.300583035f,
-    0.0135479616f,
-    -0.0275412817f,
-    -0.0034806137f,
-    0.0342865288f,
-    -0.0420582034f,
-    -0.186011374f,
-    0.11206159f,
-    0.00602817442f,
-    -0.0134207038f,
-    0.047870364f,
-    0.150576532f,
-    -0.0213764627f,
-    0.0136984214f,
-    0.00352038234f,
-    0.00102307834f,
-    -0.000959502242f,
-    0.0549549572f,
-    0.734076381f,
-    0.010336143f,
-    0.0729457736f,
-    0.0619368181f,
-    0.0166818034f,
-    -0.347239286f,
-    -0.444242895f,
-    -0.0318141505f,
-    0.0736867413f,
-    0.0786765888f,
-    0.00738393888f,
-    0.0661436692f,
-    0.219980881f,
-    -0.000330626644f,
-    -0.0323352031f,
-    -0.013685327f,
-    0.0125236576f,
-    0.20317021f,
-    -0.946965635f,
-    0.0120346425f,
-    0.0379865915f,
-    0.0434394404f,
-    -0.0137886265f,
-    -0.246061817f,
-    -0.449618071f,
-    0.0144738117f,
-    0.0677069873f,
-    0.0830139294f,
-    -0.0399225019f,
-    -0.035209462f,
-    0.153272286f,
-    0.00571329799f,
-    -0.0148759289f,
-    -0.00444148015f,
-    0.0154209146f,
-    0.126906693f,
-    -0.598590255f,
-    -0.336264938f,
-    0.249968991f,
-    -0.00146815018f,
-    0.39307344f,
-    0.329136789f,
-    0.08951471f,
-    0.180780381f,
-    0.233287409f,
-    0.533066511f,
-    0.278131902f,
-    -0.405852705f,
-    0.310305893f,
-    -0.935154378f,
-    -0.530068934f,
-    1.00221562f,
-    -0.197998837f,
-    0.0659225583f,
-    -0.456700563f,
-    -0.0219681691f,
-    -0.0786278993f,
-    0.0672209486f,
-    -0.0439613797f,
-    0.117257394f,
-    0.151135147f,
-    -0.0256771147f,
-    0.0277955774f,
-    0.123972483f,
-    -0.125673413f,
-    0.144424602f,
-    -0.148459047f,
-    0.0131278094f,
-    -0.0539841391f,
-    0.024013266f,
-    -0.00229147519f,
-    0.0492661782f,
-    -0.383196533f,
-    0.0884201601f,
-    -0.00402009999f,
-    0.0791591331f,
-    0.02627117f,
-    0.0173529889f,
-    -0.377305835f,
-    0.0278360359f,
-    0.174886584f,
-    -0.118841812f,
-    0.0596228354f,
-    0.202098757f,
-    0.201036856f,
-    0.00270011625f,
-    -0.0446279794f,
-    0.0434948877f,
-    -0.00901266653f,
-    -0.0398353487f,
-    -0.431504846f,
-    -0.0940354094f,
-    -0.088441737f,
-    -0.0911011547f,
-    -0.144156218f,
-    -0.135108754f,
-    0.210015491f,
-    -0.0706800073f,
-    -0.00796580967f,
-    -0.0071334918f,
-    0.12378037f,
-    -0.082425721f,
-    0.0593686588f,
-    -0.003034794f,
-    0.0111037968f,
-    -0.0164994635f,
-    0.00096970459f,
-    -0.0610453896f,
-    0.53365612f,
-    -0.272081703f,
-    -0.0882572308f,
-    0.11858061f,
-    0.725547254f,
-    -0.851140261f,
-    0.546806455f,
-    0.179718226f,
-    0.439941972f,
-    0.342064381f,
-    -0.103163116f,
-    0.161639407f,
-    -0.311188817f,
-    0.329252779f,
-    -0.113797337f,
-    0.253740698f,
-    0.318921447f,
-    -0.26687324f,
-    0.0829707906f,
-    -0.0358907394f,
-    -0.0717991814f,
-    0.111974038f,
-    0.129488483f,
-    -0.0320497118f,
-    0.0586829633f,
-    -0.0190763753f,
-    -0.0163752623f,
-    -0.11590533f,
-    -0.157802612f,
-    0.165030643f,
-    0.0258833747f,
-    0.141098678f,
-    -0.181780696f,
-    0.475087374f,
-    0.548292577f,
-    0.315597743f,
-    0.155289069f,
-    -0.196595609f,
-    0.351680607f,
-    -0.00146437075f,
-    -0.0257229861f,
-    -0.0119655449f,
-    -0.215780586f,
-    -0.0712371692f,
-    -0.104402132f,
-    -0.166873664f,
-    0.0165513158f,
-    0.145166457f,
-    -0.16727902f,
-    -0.0965339318f,
-    0.192882702f,
-    -0.18790029f,
-    0.402287394f,
-    0.514791965f,
-    0.270526111f,
-    -0.208685294f,
-    0.0794391856f,
-    0.191200256f,
-    0.09542384f,
-    0.089841038f,
-    0.0931917354f,
-    -0.298152655f,
-    0.063268654f,
-    -0.0527540967f,
-    -0.156194463f,
-    -0.0285386518f,
-    -0.153003231f,
-    -0.133616433f,
-    -0.200217083f,
-    -0.0568296462f,
-    0.150170907f,
-    0.0648676232f,
-    0.257570863f,
-    0.231320098f,
-    -0.176096246f,
-    0.0222893022f,
-    0.364737034f,
-    -1.22029734f,
-    -0.423109084f,
-    -0.177036077f,
-    0.379528344f,
-    -0.367580116f,
-    0.903870404f,
-    -0.817268908f,
-    0.0326061659f,
-    0.441363305f,
-    -0.808317661f,
-    -0.247720852f,
-    -0.117370494f,
-    -0.171316341f,
-    0.239265785f,
-    -0.000658903678f,
-    -0.407434732f,
-    -0.0863912031f,
-    -0.261075377f,
-    -0.165777564f,
-    -0.0755321607f,
-    0.0311758649f,
-    0.0979933739f,
-    0.136649027f,
-    0.0527689941f,
-    0.110009655f,
-    -0.231984541f,
-    -0.150287271f,
-    -0.23276037f,
-    -0.0064630718f,
-    -0.20395413f,
-    -0.197566494f,
-    -0.566420615f,
-    -0.139350429f,
-    -0.228113353f,
-    0.0187967122f,
-    -0.0344616584f,
-    0.0829470009f,
-    0.237451032f,
-    -0.145290941f,
-    0.0371040627f,
-    -0.180206016f,
-    -0.13311018f,
-    0.148776919f,
-    -0.225107476f,
-    0.017937934f,
-    -0.164880827f,
-    0.0848804787f,
-    0.111005634f,
-    -0.0585648865f,
-    0.268760353f,
-    -0.156235352f,
-    0.0508868657f,
-    0.199008629f,
-    -0.248886153f,
-    -0.0884706602f,
-    -0.354773432f,
-    -0.110232525f,
-    0.097562708f,
-    0.122192763f,
-    -0.203753442f,
-    -0.0220956448f,
-    0.300112963f,
-    0.0969282612f,
-    0.158462957f,
-    0.065214999f,
-    -0.0559996143f,
-    0.290421307f,
-    0.601416707f,
-    0.173553407f,
-    0.0679172948f,
-    0.139514253f,
-    0.639793754f,
-    0.595725f,
-    0.228672221f,
-    0.125549152f,
-    -0.417258799f,
-    0.532055259f,
-    -0.695860445f,
-    -0.22879152f,
-    -0.117811494f,
-    -0.648795903f,
-    -0.474656641f,
-    0.897781909f,
-    0.0316598788f,
-    0.263967514f,
-    0.14760375f,
-    0.0116921933f,
-    -0.0108992942f,
-    0.0787378848f,
-    -0.222190291f,
-    -0.333340466f,
-    0.350658029f,
-    0.149158806f,
-    0.0475812107f,
-    -0.0830783173f,
-    0.182674497f,
-    0.0260209367f,
-    -0.1828513f,
-    -0.0625798851f,
-    0.0349655375f,
-    0.0832568482f,
-    -0.0848430097f,
-    -0.0684579536f,
-    -0.620818794f,
-    -0.362327546f,
-    -0.0562563278f,
-    -0.348935723f,
-    -0.920360446f,
-    0.00371341989f,
-    -0.0318348147f,
-    0.0482412651f,
-    0.0100648198f,
-    -0.0321812741f,
-    0.0397563949f,
-    0.276662439f,
-    -0.025499817f,
-    -0.344580054f,
-    0.52697885f,
-    -0.446119219f,
-    0.263266534f,
-    0.048449818f,
-    -0.0985448211f,
-    -0.0217188783f,
-    0.12052349f,
-    0.193878576f,
-    0.0771778896f,
-    -0.328556567f,
-    -0.917501986f,
-    0.0528322235f,
-    -0.111473925f,
-    0.288629562f,
-    0.203885779f,
-    -0.107121781f,
-    -0.0836843476f,
-    -0.0972784981f,
-    0.136616021f,
-    -0.039855063f,
-    -0.112426169f,
-    -0.0301022828f,
-    -0.544496536f,
-    -0.0955949724f,
-    -0.25732556f,
-    0.0821035877f,
-    -0.551635444f,
-    -0.697915494f,
-    0.400376856f,
-    -0.144447535f,
-    0.609956026f,
-    0.449059546f,
-    0.220692649f,
-    0.00563674048f,
-    -0.51202476f,
-    0.708183706f,
-    -0.0823447928f,
-    -0.52478379f,
-    0.037751887f,
-    0.183620006f,
-    0.500270188f,
-    -0.155276731f,
-    -0.0432680361f,
-    -0.0779945627f,
-    -0.0901541188f,
-    -0.111379944f,
-    -0.0765521303f,
-    -0.180463538f,
-    -0.0194951333f,
-    0.0680495128f,
-    -0.0954369828f,
-    -0.0314615555f,
-    -0.144507036f,
-    -0.0388588943f,
-    -0.0209448095f,
-    -0.102267414f,
-    0.079779163f,
-    0.0448272824f,
-    0.0419289656f,
-    0.0179472491f,
-    0.00780927157f,
-    0.0243407451f,
-    0.047702156f,
-    0.0957163274f,
-    0.0380610749f,
-    -0.0249969903f,
-    0.00515407091f,
-    0.0797748268f,
-    0.0851003304f,
-    0.220019922f,
-    -0.357321113f,
-    0.026951436f,
-    0.0118372329f,
-    -0.00855999812f,
-    -0.0236534085f,
-    -0.0683448985f,
-    -0.835489154f,
-    -0.0322686173f,
-    -0.0273920316f,
-    -0.0291251261f,
-    -0.0943413377f,
-    -0.213537559f,
-    -0.181599751f,
-    -0.232553259f,
-    0.145303473f,
-    0.037667267f,
-    -0.24855116f,
-    -0.24051255f,
-    -0.428465217f,
-    -0.00799547881f,
-    -0.246761441f,
-    -0.136331767f,
-    -0.0215395726f,
-    -0.181716904f,
-    0.206022084f,
-    0.10649088f,
-    -0.194210947f,
-    0.20393306f,
-    0.669626772f,
-    0.722047806f,
-    0.093189016f,
-    -0.461152285f,
-    0.0128888162f,
-    0.439610511f,
-    0.415558159f,
-    0.431113631f,
-    0.514324248f,
-    0.103293501f,
-    0.684579253f,
-    -0.33730641f,
-    0.0934910104f,
-    0.605639756f,
-    -0.416589171f,
-    0.0518259481f,
-    0.104484469f,
-    0.105649821f,
-    0.13106215f,
-    0.125276625f,
-    0.144136131f,
-    -0.0498236306f,
-    -0.0493632294f,
-    0.0280261915f,
-    -0.0208750535f,
-    0.0672143325f,
-    0.19184497f,
-    -0.0157077461f,
-    -0.0282561742f,
-    -0.0950072706f,
-    0.0335669927f,
-    -0.0895609856f,
-    -0.233938009f,
-    0.128741294f,
-    0.21595487f,
-    0.169179395f,
-    0.110438764f,
-    -0.00270493422f,
-    -0.21807155f,
-    0.319750041f,
-    0.267321676f,
-    0.250353456f,
-    0.4375377f,
-    0.0880190581f,
-    0.213034362f,
-    -0.137480006f,
-    -0.193822399f,
-    -0.39286235f,
-    -0.0687411129f,
-    -0.0499877296f,
-    0.204800203f,
-    -0.101521052f,
-    -0.184072226f,
-    -0.0134938294f,
-    -0.0438117757f,
-    0.0730600134f,
-    -0.267662138f,
-    0.234730467f,
-    0.0887582451f,
-    0.0822641775f,
-    0.0386104025f,
-    -0.000654937467f,
-    0.139305145f,
-    -0.157022357f,
-    -0.0528170988f,
-    0.130827248f,
-    -0.00133016577f,
-    -0.00777009735f,
-    0.149428561f,
-    -0.481059849f,
-    0.0794200152f,
-    -0.0317041911f,
-    0.813333631f,
-    0.156108677f,
-    0.669233859f,
-    -0.27749455f,
-    0.0914489105f,
-    0.159320042f,
-    -0.546823502f,
-    0.553968251f,
-    0.272330135f,
-    0.401908129f,
-    0.00528037967f,
-    0.307021201f,
-    -0.518522382f,
-    0.334646374f,
-    -0.249633417f,
-    0.0535853282f,
-    -0.119376205f,
-    0.0505777821f,
-    -0.324623406f,
-    -0.147153407f,
-    0.067020312f,
-    0.193799406f,
-    -0.130087122f,
-    0.184588075f,
-    0.0595951714f,
-    -0.0701579526f,
-    -0.105733737f,
-    -0.133412033f,
-    -0.192366898f,
-    -0.222749338f,
-    -0.00815801881f,
-    0.0786651894f,
-    -0.001446718f,
-    0.136215404f,
-    -0.112398215f,
-    -0.153114632f,
-    -0.258102208f,
-    -0.109383807f,
-    -0.0643849075f,
-    -0.102717608f,
-    -0.0599655136f,
-    0.0159161389f,
-    -0.00965123996f,
-    -0.185906455f,
-    0.114197113f,
-    0.0309348609f,
-    -0.067304723f,
-    0.120613463f,
-    0.0195027292f,
-    -0.108209632f,
-    0.099880591f,
-    -0.0359357484f,
-    -0.0510859005f,
-    -0.22176303f,
-    0.0795331895f,
-    -0.22134918f,
-    0.627900779f,
-    -0.164595157f,
-    -0.0639208704f,
-    -0.192175359f,
-    0.359550744f,
-    0.0727395117f,
-    0.685350537f,
-    0.190056801f,
-    0.169738516f,
-    0.320223659f,
-    0.0365936235f,
-    0.10938856f,
-    0.179485604f,
-    0.028861722f,
-    0.121144861f,
-    0.444255978f,
-    0.213994727f,
-    0.551053345f,
-    0.12976037f,
-    -0.32904166f,
-    -0.0536703169f,
-    -0.183060005f,
-    0.120761834f,
-    -0.273113191f,
-    -0.193261489f,
-    0.0109263118f,
-    -0.058256872f,
-    0.0490219742f,
-    -0.317464113f,
-    -0.594732165f,
-    0.222740978f,
-    -0.0488459133f,
-    -0.0671636462f,
-    -0.0436842069f,
-    -0.0845794976f,
-    -0.0374646932f,
-    -0.35784626f,
-    -0.210188717f,
-    -0.13194342f,
-    -0.1539682f,
-    -0.104031667f,
-    -0.29204905f,
-    -0.148935348f,
-    0.093869552f,
-    0.0873055011f,
-    0.0683833286f,
-    0.0816393197f,
-    0.144601896f,
-    -0.714668095f,
-    0.0614345223f,
-    -0.0382773764f,
-    0.0655549914f,
-    -0.0615131855f,
-    0.0233947895f,
-    -0.165001526f,
-    0.0723389536f,
-    -0.038349092f,
-    -0.13368623f,
-    -0.0488641448f,
-    -0.161910534f,
-    -0.14366889f,
-    -0.022888083f,
-    -0.0333589166f,
-    0.0953429937f,
-    -0.0947889239f,
-    0.0908548385f,
-    -0.0505312905f,
-    -0.109693967f,
-    0.117231764f,
-    -0.0231866259f,
-    0.235611051f,
-    -0.05812978f,
-    -0.13047944f,
-    -0.0981884822f,
-    0.201866224f,
-    0.0988980383f,
-    0.203553647f,
-    -0.0420718193f,
-    0.153427348f,
-    0.0591607057f,
-    -0.0848791972f,
-    -0.0330034122f,
-    0.076174587f,
-    -0.0753726214f,
-    0.256996691f,
-    0.39395225f,
-    0.222272843f,
-    -0.0398086868f,
-    -0.924315155f,
-    0.699756801f,
-    -1.17936599f,
-    0.557080865f,
-    -0.24136968f,
-    0.275698572f,
-    -0.501634538f,
-    -0.258966684f,
-    0.210878953f,
-    -0.492018729f,
-    0.166261435f,
-    0.004552837f,
-    -0.307355583f,
-    -0.318138778f,
-    -0.645166934f,
-    -0.114447311f,
-    -0.0671354681f,
-    -0.0742307827f,
-    0.0601832345f,
-    -0.12114314f,
-    -0.157649472f,
-    -0.0917536095f,
-    -0.222220317f,
-    0.0337064117f,
-    -0.0597106256f,
-    -0.00932833645f,
-    0.00158322847f,
-    -0.120634407f,
-    -0.127468705f,
-    -0.142555922f,
-    0.240737841f,
-    0.0453580394f,
-    0.0327029563f,
-    -0.288416773f,
-    -0.00896670297f,
-    0.186482385f,
-    0.0340303071f,
-    0.117670052f,
-    0.145555034f,
-    -0.153254211f,
-    -0.0306744203f,
-    -0.0519810133f,
-    -0.290757447f,
-    0.0533552542f,
-    -0.0227261186f,
-    0.0579988398f,
-    0.130642012f,
-    0.0282798186f,
-    -0.105240606f,
-    0.0858023837f,
-    0.141845316f,
-    0.00381381204f,
-    -0.0816450268f,
-    0.0546937585f,
-    -0.0566968806f,
-    -0.0807035044f,
-    -0.0511513241f,
-    -0.221844926f,
-    0.0624566711f,
-    0.101994552f,
-    -0.103440195f,
-    -0.0720308796f,
-    -0.0377379954f,
-    -0.120321706f,
-    -0.0418350473f,
-    -0.115037143f,
-    0.00228702696f,
-    0.144421399f,
-    -0.138549477f,
-    -0.200199604f,
-    -0.012018769f,
-    -0.169050321f,
-    -0.0488194115f,
-    0.0795217082f,
-    0.0660576969f,
-    -0.104960866f,
-    0.0398860872f,
-    0.0191672258f,
-    0.427314222f,
-    0.314845502f,
-    -0.0616263151f,
-    0.192402124f,
-    0.198853627f,
-    0.090693973f,
-    0.194811061f,
-    0.00586435199f,
-    0.117237873f,
-    0.0304926299f,
-    -0.330035776f,
-    -0.108412236f,
-    -0.0648701191f,
-    0.230319574f,
-    -0.0682587177f,
-    0.229308784f,
-    0.299034834f,
-    -0.0640124232f,
-    -0.0762560442f,
-    0.119224764f,
-    0.0218221303f,
-    -0.0966265649f,
-    -0.199867129f,
-    0.0267505851f,
-    0.0419577695f,
-    0.0784355551f,
-    0.160422504f,
-    0.0391016155f,
-    0.0572678708f,
-    0.0250213984f,
-    0.0710889176f,
-    0.0663850829f,
-    -0.145608038f,
-    0.0427745543f,
-    -0.229348212f,
-    0.0214061365f,
-    0.102577411f,
-    0.105400003f,
-    -0.103882812f,
-    0.098611787f,
-    -0.114075288f,
-    0.011778187f,
-    0.122762904f,
-    0.0151383039f,
-    -0.277138978f,
-    -0.130039826f,
-    -0.265268534f,
-    -0.143608436f,
-    -0.540323496f,
-    -0.0348086841f,
-    -0.154655039f,
-    -0.190752074f,
-    -0.110869601f,
-    0.0564384498f,
-    -0.405536473f,
-    0.11008057f,
-    -0.167652085f,
-    0.0451175161f,
-    -0.0537001044f,
-    -0.550586462f,
-    -0.43535322f,
-    -0.368644267f,
-    -0.506221116f,
-    -0.415773362f,
-    0.0122666191f,
-    0.0674380884f,
-    0.151497081f,
-    0.0923110247f,
-    0.323450565f,
-    0.023222087f,
-    -0.0297017097f,
-    -0.0904677734f,
-    0.104964003f,
-    -0.0260316152f,
-    0.326415032f,
-    0.789115429f,
-    0.495272309f,
-    0.441865176f,
-    -0.579993546f,
-    0.552707434f,
-    -0.455564857f,
-    0.33328712f,
-    0.195112124f,
-    -0.685172439f,
-    0.003143657f,
-    -0.274307728f,
-    0.331762344f,
-    -0.267942041f,
-    -0.0343040526f,
-    0.0235557742f,
-    -0.176633209f,
-    -0.00182777143f,
-    -0.120159514f,
-    -0.165586516f,
-    -0.0818145052f,
-    -0.165402338f,
-    -0.0233940259f,
-    -0.045715712f,
-    -0.163899034f,
-    -0.0252915528f,
-    0.0834722295f,
-    -0.104314975f,
-    0.203057945f,
-    0.156299815f,
-    -0.326268375f,
-    -0.124628641f,
-    -0.0270725191f,
-    0.0415703319f,
-    -0.0961005762f,
-    -0.0417908132f,
-    0.0309348032f,
-    -0.0229402706f,
-    -0.0182094499f,
-    -0.00509249512f,
-    -0.0152613586f,
-    -0.151194394f,
-    0.211405322f,
-    0.0794121027f,
-    -0.0750945508f,
-    -0.137154669f,
-    -0.234895185f,
-    -0.138213292f,
-    -0.0529647283f,
-    -0.107116349f,
-    0.0655621737f,
-    0.103434838f,
-    -0.0749790967f,
-    0.337472528f,
-    0.128928408f,
-    0.0827406123f,
-    -0.03742145f,
-    0.178835794f,
-    -0.710225701f,
-    0.158309564f,
-    -0.342295706f,
-    0.00116564182f,
-    0.0118570086f,
-    -0.0546392575f,
-    -0.0105886022f,
-    -0.176536977f,
-    0.0321256593f,
-    -0.0768093839f,
-    -0.0208240245f,
-    -0.0145562468f,
-    -0.206373319f,
-    -0.291801572f,
-    -0.190143228f,
-    -0.0440782458f,
-    -0.308749348f,
-    -0.0481246486f,
-    0.115309782f,
-    -0.0898751691f,
-    -0.0535585284f,
-    -0.0869345441f,
-    -0.16089727f,
-    0.086908862f,
-    0.113670968f,
-    -0.174988478f,
-    -0.0768239349f,
-    -0.047540497f,
-    -0.0600021519f,
-    -0.146179467f,
-    0.332772791f,
-    0.244854525f,
-    -0.181740671f,
-    -0.0124077275f,
-    -0.280300528f,
-    -0.101044931f,
-    0.00426148903f,
-    -0.172911361f,
-    -0.0735776946f,
-    -0.198563352f,
-    -0.256841689f,
-    0.353101522f,
-    0.0279062744f,
-    -0.133012712f,
-    -0.0437886789f,
-    0.588336885f,
-    1.35203052f,
-    -0.0862999931f,
-    -0.256041974f,
-    -0.217727065f,
-    -0.152908206f,
-    -0.0550493039f,
-    -0.0204031281f,
-    0.0910874605f,
-    -0.118821986f,
-    0.0536605641f,
-    0.0500407591f,
-    -0.101197101f,
-    0.00670321239f,
-    0.0335346721f,
-    0.0242832657f,
-    0.0238263849f,
-    0.0586579405f,
-    0.130556762f,
-    -0.00797482394f,
-    0.129874006f,
-    0.0166296232f,
-    0.0770444423f,
-    0.0448188037f,
-    0.095067203f,
-    -0.0906836838f,
-    0.260206282f,
-    -0.0185685623f,
-    0.164078593f,
-    0.135952264f,
-    0.0469171293f,
-    -0.317476332f,
-    0.239058167f,
-    0.0692567304f,
-    0.179726392f,
-    0.0479587577f,
-    0.107986666f,
-    -4.52961503e-05f,
-    0.0315949358f,
-    0.179472312f,
-    0.139305294f,
-    0.0989076048f,
-    0.195278168f,
-    -0.0499197133f,
-    0.0179640483f,
-    0.175898224f,
-    -0.357745469f,
-    -0.185439005f,
-    0.630868971f,
-    0.316389203f,
-    0.0371670425f,
-    0.0956562012f,
-    -0.156396866f,
-    0.4211393f,
-    -0.30426085f,
-    -0.10838183f,
-    -0.19902195f,
-    0.202597484f,
-    -0.243825972f,
-    0.401519984f,
-    0.543762743f,
-    -0.592841804f,
-    0.472569644f,
-    -0.441522002f,
-    0.149017945f,
-    -0.343037099f,
-    -0.256952792f,
-    -0.141287163f,
-    -0.00775221689f,
-    0.0990487859f,
-    -0.00690405862f,
-    0.0113994461f,
-    -0.0598971881f,
-    -0.170097858f,
-    0.168442577f,
-    -0.142957792f,
-    0.0735734254f,
-    -0.131490931f,
-    -0.35617277f,
-    0.0718475655f,
-    -0.241283089f,
-    0.204179063f,
-    0.311081916f,
-    0.142517149f,
-    0.168676451f,
-    -0.119859546f,
-    0.290017873f,
-    -0.160168454f,
-    -0.153524637f,
-    -0.0947197303f,
-    0.28523019f,
-    0.345079899f,
-    0.530938625f,
-    0.254538447f,
-    -0.177891463f,
-    -0.192460924f,
-    0.0314176939f,
-    -0.238392338f,
-    -0.231304273f,
-    -0.00142887002f,
-    -0.198986918f,
-    0.178607136f,
-    -0.0474961065f,
-    0.0552055202f,
-    -0.0708224699f,
-    0.00527971936f,
-    -0.0731117725f,
-    0.477975577f,
-    0.0443970785f,
-    -0.196957141f,
-    0.134686768f,
-    0.0762641877f,
-    0.260251641f,
-    -0.0798139125f,
-    0.0881259739f,
-    0.306648731f,
-    -0.104396045f,
-    -0.0770304203f,
-    -0.136753812f,
-    -0.17658408f,
-    0.532241523f,
-    0.50030905f,
-    -0.0536105558f,
-    0.422902435f,
-    1.01389229f,
-    -0.0481119752f,
-    -0.434554726f,
-    0.369995713f,
-    -0.0877772719f,
-    0.418274015f,
-    -0.490831584f,
-    -0.601176143f,
-    -0.592036903f,
-    0.834720075f,
-    0.445466459f,
-    -0.115680948f,
-    0.386708796f,
-    0.051833868f,
-    -0.655744791f,
-    -0.0762274414f,
-    0.310995162f,
-    0.0158713982f,
-    -0.113714591f,
-    -0.140238971f,
-    0.324495465f,
-    0.252955198f,
-    0.0556549132f,
-    -0.26186493f,
-    0.193261713f,
-    0.294905901f,
-    -0.243056893f,
-    -0.495643109f,
-    0.587792814f,
-    0.165307358f,
-    -0.045893345f,
-    0.0474878214f,
-    -0.0314576589f,
-    -0.406758487f,
-    0.133858204f,
-    0.169137374f,
-    0.323978037f,
-    0.248654529f,
-    -0.258320123f,
-    -0.121062852f,
-    -0.0590989292f,
-    -0.179658726f,
-    0.207047552f,
-    0.225881666f,
-    -0.0707306564f,
-    -0.184742108f,
-    -0.00721999863f,
-    0.0864961073f,
-    -0.0878195316f,
-    0.155501291f,
-    -0.0889338404f,
-    -0.0684351325f,
-    -0.15320687f,
-    -0.523227215f,
-    -0.115148969f,
-    0.166544884f,
-    0.627472997f,
-    -0.29913187f,
-    -0.240128681f,
-    0.0340427756f,
-    0.392243057f,
-    0.249277532f,
-    -0.224078983f,
-    0.07031703f,
-    0.370228618f,
-    -0.264994234f,
-    -0.513675809f,
-    -0.197438553f,
-    0.432436138f,
-    -0.0695493147f,
-    0.199592769f,
-    0.35958147f,
-    -0.036837589f,
-    -0.356063694f,
-    -0.162454411f,
-    0.0193438269f,
-    0.187823504f,
-    -0.678046942f,
-    -0.290522218f,
-    -0.234926417f,
-    0.303240418f,
-    -0.36501205f,
-    -0.244342014f,
-    -0.0600830466f,
-    -0.0948337913f,
-    -0.293538272f,
-    0.0308160093f,
-    -0.200693935f,
-    -0.27982527f,
-    0.22725378f,
-    -0.094583936f,
-    -0.132877007f,
-    -0.0216341466f,
-    -0.270980686f,
-    0.032521788f,
-    0.270779282f,
-    0.320024043f,
-    0.136084378f,
-    0.0376128256f,
-    -0.210607708f,
-    0.0673111603f,
-    -0.236148655f,
-    -0.137558788f,
-    0.241716981f,
-    -0.117170468f,
-    -0.0182848051f,
-    0.117692068f,
-    0.00916733686f,
-    0.0678152218f,
-    0.0687659755f,
-    0.0915961117f,
-    0.307116449f,
-    0.1931272f,
-    0.195927873f,
-    0.0330154076f,
-    0.0826165602f,
-    0.065019384f,
-    0.146117926f,
-    -0.20774138f,
-    0.0654244348f,
-    0.0511493459f,
-    -0.0497630276f,
-    -0.182122409f,
-    -0.0746433586f,
-    0.00679971743f,
-    -0.102891944f,
-    -0.125131786f,
-    0.0103310551f,
-    0.0348905027f,
-    0.156693012f,
-    -0.0307797249f,
-    0.0595415346f,
-    -0.0310573913f,
-    0.0789554268f,
-    0.271517068f,
-    0.135436282f,
-    0.149715379f,
-    0.0360571668f,
-    0.234488919f,
-    -0.301850975f,
-    0.34194386f,
-    -0.0629542023f,
-    0.172238305f,
-    0.103567988f,
-    -0.12782234f,
-    0.136975124f,
-    0.127197012f,
-    0.420429081f,
-    -0.243052244f,
-    -0.167192921f,
-    0.485649973f,
-    0.319841087f,
-    -0.266684294f,
-    0.425268888f,
-    0.750342667f,
-    0.125548571f,
-    -0.0191239789f,
-    -0.204060838f,
-    -0.0428533182f,
-    0.165749222f,
-    0.143341288f,
-    -0.269152462f,
-    -0.372700065f,
-    -0.319214463f,
-    0.0101419874f,
-    0.0176876709f,
-    -0.162041694f,
-    -0.0677598044f,
-    -0.0143079273f,
-    -0.346304655f,
-    -0.22881268f,
-    -0.417520434f,
-    -0.29699865f,
-    -0.260143518f,
-    -0.449609041f,
-    0.369638264f,
-    -0.312429875f,
-    -0.261035711f,
-    -0.173840657f,
-    0.00344708445f,
-    0.303478926f,
-    -0.351835489f,
-    -0.0406893529f,
-    -0.138836637f,
-    0.0902438611f,
-    -0.15125905f,
-    -0.298170984f,
-    0.438161105f,
-    -0.175641239f,
-    -0.143300161f,
-    -0.318568647f,
-    0.0339598656f,
-    0.273231715f,
-    0.387224495f,
-    -0.0649677664f,
-    0.0306851771f,
-    -0.0773795843f,
-    -0.0771140978f,
-    0.138683453f,
-    0.1361029f,
-    0.1763338f,
-    0.138067514f,
-    0.173025981f,
-    0.139558926f,
-    0.121444225f,
-    0.151194632f,
-    0.344341725f,
-    0.193694562f,
-    0.253284872f,
-    -0.089330852f,
-    0.064525269f,
-    -0.139717877f,
-    0.180656567f,
-    0.276411742f,
-    0.067350626f,
-    -0.463683903f,
-    0.336487055f,
-    -0.465005606f,
-    0.495838314f,
-    0.178243309f,
-    0.191439882f,
-    -0.265802145f,
-    -0.134343073f,
-    -0.0527970567f,
-    -0.448106557f,
-    -0.331229746f,
-    0.461991757f,
-    -0.52609843f,
-    -0.738187909f,
-    -0.583274186f,
-    -0.0461513437f,
-    0.0321800224f,
-    -0.211429894f,
-    -0.00930533651f,
-    0.399375409f,
-    -0.00869760569f,
-    -0.0309554338f,
-    0.110588282f,
-    0.081680581f,
-    -0.117330305f,
-    -0.251681924f,
-    -0.247767225f,
-    -0.0307760965f,
-    -0.0263815559f,
-    -0.0192796346f,
-    -0.0831981227f,
-    -0.497095853f,
-    0.112677649f,
-    -0.0634514913f,
-    -0.0946953148f,
-    -0.177060083f,
-    -0.206850767f,
-    0.230848193f,
-    -0.193116277f,
-    -0.0641803667f,
-    -0.18318066f,
-    -0.210208952f,
-    -0.213836953f,
-    -0.201510355f,
-    0.257652789f,
-    0.444827735f,
-    -0.0939728022f,
-    -0.0293442588f,
-    -0.143569499f,
-    0.0395009592f,
-    -0.232463449f,
-    0.0267518796f,
-    0.152513415f,
-    0.0436136685f,
-    0.057627663f,
-    0.105509833f,
-    0.135316744f,
-    0.0927302465f,
-    0.0855649859f,
-    0.26530537f,
-    -0.0469772965f,
-    -0.138744369f,
-    0.103846744f,
-    0.0257813763f,
-    -0.0843293443f,
-    -0.166094452f,
-    -0.0439791195f,
-    -0.0701237097f,
-    0.0754311159f,
-    -0.063143827f,
-    0.536367953f,
-    -0.214776337f,
-    -0.0302201901f,
-    0.127061322f,
-    -0.143678561f,
-    -0.0551943108f,
-    0.772420406f,
-    0.138803333f,
-    0.288101226f,
-    -0.919005334f,
-    0.0185176525f,
-    0.00935063232f,
-    0.243831888f,
-    -0.39196986f,
-    -0.081883423f,
-    -0.523019373f,
-    -0.553617597f,
-    0.116791144f,
-    0.140534714f,
-    0.0557657816f,
-    0.111191675f,
-    0.024996018f,
-    0.103502877f,
-    0.0413700268f,
-    0.0109461173f,
-    0.0802151114f,
-    0.220318943f,
-    0.0821969286f,
-    0.12996906f,
-    -0.0560302958f,
-    -0.149437025f,
-    -0.0224109702f,
-    0.143744737f,
-    -0.127283305f,
-    0.154110014f,
-    -0.126627713f,
-    0.0791257992f,
-    0.185230821f,
-    0.0284365937f,
-    -0.168269128f,
-    -0.258627445f,
-    -0.0533706509f,
-    -0.240679532f,
-    -0.111877583f,
-    0.196009234f,
-    -0.233653009f,
-    0.0213059932f,
-    -0.105173536f,
-    -0.129015312f,
-    0.0862896144f,
-    0.25685358f,
-    -0.14850536f,
-    -0.119752161f,
-    0.0920667723f,
-    -0.0935089737f,
-    -0.0551220141f,
-    0.0733904615f,
-    0.244723693f,
-    0.102900714f,
-    -0.15775007f,
-    0.0613219887f,
-    -0.153664708f,
-    -0.349462211f,
-    -0.0618619993f,
-    -0.0664012209f,
-    0.162872925f,
-    0.0306008756f,
-    -0.142181978f,
-    -0.235186473f,
-    0.357436657f,
-    -0.22828193f,
-    0.00800762977f,
-    0.40609628f,
-    0.319911301f,
-    -0.0147775933f,
-    0.659579873f,
-    -0.233283713f,
-    -0.202653721f,
-    -0.319093287f,
-    -0.744725883f,
-    -0.0963212699f,
-    0.264960319f,
-    -0.424700111f,
-    -0.224975511f,
-    -0.722756147f,
-    0.162389114f,
-    -0.0111915329f,
-    0.459106117f,
-    0.130714536f,
-    0.0291007329f,
-    -0.0898427442f,
-    -0.0563552044f,
-    -0.00690592406f,
-    0.133364797f,
-    0.0930068269f,
-    -0.0166413337f,
-    -0.0340050831f,
-    0.0148325674f,
-    -0.137482285f,
-    0.0818665698f,
-    0.220479816f,
-    0.0345323421f,
-    0.0427586697f,
-    0.141208991f,
-    -0.236098081f,
-    0.521645784f,
-    0.125019297f,
-    0.0209487788f,
-    0.0132382466f,
-    0.0173130129f,
-    0.198263511f,
-    -0.0735999569f,
-    -0.026108237f,
-    -0.0103781875f,
-    -0.0202101227f,
-    0.0143875303f,
-    -0.0637853891f,
-    0.144778788f,
-    0.0658948943f,
-    0.00489846058f,
-    -0.0262471996f,
-    0.018627936f,
-    -0.111215979f,
-    0.170737758f,
-    0.122505762f,
-    -0.0431824028f,
-    0.0870481208f,
-    0.0588259399f,
-    -0.0754437819f,
-    -0.101210341f,
-    -0.0841763243f,
-    -0.00516707683f,
-    0.0215253178f,
-    0.0357166566f,
-    0.131340429f,
-    -0.0576243699f,
-    -0.276282996f,
-    -0.0256422181f,
-    -0.0294179283f,
-    -0.117724083f,
-    0.186967701f,
-    -0.580237985f,
-    -0.0351368561f,
-    0.0475606732f,
-    0.0618246235f,
-    0.193020031f,
-    0.0606427081f,
-    0.196171314f,
-    -0.119147539f,
-    0.416209847f,
-    -0.00858832337f,
-    0.532788813f,
-    0.116121583f,
-    0.186319411f,
-    -0.0747691467f,
-    -0.322490066f,
-    -0.389920413f,
-    -0.369598299f,
-    -0.345728785f,
-    0.494861633f,
-    -0.35605675f,
-    -0.0136896893f,
-    0.0652199164f,
-    0.0219608564f,
-    0.0556277707f,
-    -0.0315487497f,
-    -0.0499449857f,
-    -0.170733973f,
-    -0.156915992f,
-    -0.204194263f,
-    -0.164826855f,
-    -0.149145544f,
-    -0.0899075195f,
-    0.0764857903f,
-    0.107416995f,
-    0.0985934213f,
-    0.0391226225f,
-    -0.122678451f,
-    -0.130244628f,
-    -0.108455546f,
-    -0.0984326899f,
-    -0.103473619f,
-    -0.117170118f,
-    0.023315357f,
-    0.0249001533f,
-    -0.00559251849f,
-    0.0838891044f,
-    0.13185218f,
-    0.0719638094f,
-    -0.0932370573f,
-    -0.137218669f,
-    -0.0640774295f,
-    -0.0535033345f,
-    0.0534853339f,
-    0.0787431896f,
-    0.0554250814f,
-    0.180693522f,
-    0.177916437f,
-    0.178504825f,
-    0.034248665f,
-    0.0904801786f,
-    0.032926511f,
-    0.13449645f,
-    -0.00183142477f,
-    0.0296126753f,
-    -0.0415589996f,
-    0.00752302771f,
-    0.0102339545f,
-    -0.00414298987f,
-    -0.00282973214f,
-    0.00139734185f
-};
-
-
-// -----------------------------------------------------------------------------
-// Aggressive balanced revision 5: compact code + fully packed hot weights
-// -----------------------------------------------------------------------------
-// v4 proved that packed convolution weights recover a large amount of CPU.
-// v5 restores the remaining small hot parameter groups from v2 as packed RAM:
-//   - rechannel[3]
-//   - convW[]
-//   - l1x1W[]
-//   - per-layer convB / mixinW / l1x1B / preCurrent
-//   - headW / headB / headScale
-//
-// This spends about 4 KB more RAM than v4, but avoids repeated raw-model
-// address arithmetic and raw-layout loads inside the callback. It keeps the
-// compact v2/v4 loop kernels and constexpr topology, so .text should remain
-// much smaller than the original fully unrolled header.
-
-inline constexpr int kLayerCols[kNumLayers] = {
-    7, 17, 37, 87, 207, 507, 1197, 7, 17, 37, 87, 207, 507, 1197,
-    16, 184, 7, 17, 37, 87, 207, 507, 1197
-};
-
-// Float offsets into A2State::history.
-inline constexpr int kLayerHistoryOffset[kNumLayers] = {
-    0, 21, 72, 183, 444, 1065, 2586, 6177, 6198, 6249,
-    6360, 6621, 7242, 8763, 12354, 12402, 12954, 12975, 13026, 13137,
-    13398, 14019, 15540
-};
-
-// Float offsets into A2SharedWeights::convW.
-inline constexpr int kLayerConvOffset[kNumLayers] = {
-    0, 54, 108, 162, 216, 270, 324, 378, 432, 486, 540, 594,
-    648, 702, 756, 891, 1026, 1080, 1134, 1188, 1242, 1296, 1350
-};
-
-// Float offsets into A2SharedWeights::l1x1W.
-inline constexpr int kLayerL1Offset[kNumLayers] = {
-    0, 9, 18, 27, 36, 45, 54, 63, 72, 81, 90, 99,
-    108, 117, 126, 135, 144, 153, 162, 171, 180, 189, 198
-};
-
-struct A2LayerRuntime
-{
-    // Keep only the per-layer values that are actually model parameters.
-    NAM_A2_ALIGN32 float convB[3];
-    NAM_A2_ALIGN32 float mixinW[3];
-    NAM_A2_ALIGN32 float preCurrent[3];
-    NAM_A2_ALIGN32 float l1x1B[3];
-};
-
-
-// Raw model layout helpers.
-// Layout:
-//   rechannel[3]
-//   per layer:
-//      conv[out][in][tap]
-//      convB[3]
-//      mixinW[3]
-//      l1x1[out][in]
-//      l1x1B[3]
-//   headW[in][tap]
-//   headB
-//   headScale
-inline int raw_layer_base(int li) noexcept
-{
-    int off = 3;
-    for(int i = 0; i < li; ++i)
-        off += kKernelSizes[i] * 9 + 18;
-    return off;
-}
-
-static constexpr int kRawHeadOffset  = 3 + kConvWeightCount + kNumLayers * 18; // 1821
-static constexpr int kRawHeadBOffset = kRawHeadOffset + kHeadKernel * kChannels; // 1869
-static constexpr int kRawScaleOffset = kRawHeadBOffset + 1; // 1870
-
-struct A2SharedWeights
-{
-    NAM_A2_ALIGN32 float rechannel[3];
-    NAM_A2_ALIGN32 float convW[kConvWeightCount];       // tap-major, 3x3 per tap
-    NAM_A2_ALIGN32 float l1x1W[kLayer1x1WeightCount];   // layer-major, column-major 3x3
-    NAM_A2_ALIGN32 float headW[kHeadKernel][3];
-    float headB;
-    float headScale;
-    A2LayerRuntime layer[kNumLayers];
-    bool loaded;
-};
-
-struct A2State
-{
-    // Large dilation history remains in ordinary SRAM.
-    NAM_A2_STATE_DATA NAM_A2_ALIGN32 float history[kHistoryFloats];
-    int layerWritePos[kNumLayers];
-};
-
-struct A2HotState
-{
-    // Small, high-traffic block buffers live in DTCM.
-    NAM_A2_ALIGN32 float headHistory[kHeadKernel * 3];
-    NAM_A2_ALIGN32 float bufA[kBlockSize * 3];
-    NAM_A2_ALIGN32 float bufB[kBlockSize * 3];
-    NAM_A2_ALIGN32 float headSum[kBlockSize * 3];
-    int headWritePos;
-};
-
-// Branchless leaky-ReLU variant.
-//
-// leaky(x) = x                 for x >= 0
-//          = kLeakySlope * x   for x < 0
-//
-// Equivalent identity:
-//   leaky(x) = 0.5 * ((1 + a) * x + (1 - a) * abs(x))
-//
-// On Cortex-M7/GCC, __builtin_fabsf should lower to VABS.F32 instead of a
-// library call. This avoids a compare/branch or VFP condition transfer in the
-// post/residual path. If it regresses, set NAM_A2_BRANCHLESS_LEAKY to 0.
-#ifndef NAM_A2_BRANCHLESS_LEAKY
-#define NAM_A2_BRANCHLESS_LEAKY 1
-#endif
-
-NAM_A2_TINY_INLINE float leaky(float x) noexcept
-{
-#if NAM_A2_BRANCHLESS_LEAKY
-    constexpr float p = 0.5f * (1.0f + kLeakySlope);
-    constexpr float q = 0.5f * (1.0f - kLeakySlope);
-    return p * x + q * __builtin_fabsf(x);
-#else
-    return x >= 0.0f ? x : x * kLeakySlope;
-#endif
-}
-
-NAM_A2_TINY_INLINE void acc_tap3(const float* __restrict__ w,
-                                  const float* __restrict__ s,
-                                  float& z0,
-                                  float& z1,
-                                  float& z2) noexcept
-{
-    const float s0 = s[0];
-    const float s1 = s[1];
-    const float s2 = s[2];
-
-    z0 += s0 * w[0] + s1 * w[3] + s2 * w[6];
-    z1 += s0 * w[1] + s1 * w[4] + s2 * w[7];
-    z2 += s0 * w[2] + s1 * w[5] + s2 * w[8];
-}
-
-inline bool load_weights(A2SharedWeights& dst, const float* src, std::size_t count) noexcept
-{
-    if (src == nullptr || count != static_cast<std::size_t>(kA2WeightCount))
-        return false;
-
-    std::memset(&dst, 0, sizeof(dst));
-
-    std::size_t p = 0;
-    auto take = [&]() noexcept -> float {
-        return src[p++];
-    };
-
-    for (int c = 0; c < 3; ++c)
-        dst.rechannel[c] = take();
-
-    int convOff = 0;
-    int l1Off = 0;
-
-    for (int li = 0; li < kNumLayers; ++li)
-    {
-        A2LayerRuntime& L = dst.layer[li];
-        const int kernelSize = kKernelSizes[li];
-
-        // NAM Conv1D order: out, in, tap. Store tap-major, column-major per tap.
-        for (int out = 0; out < 3; ++out)
-            for (int in = 0; in < 3; ++in)
-                for (int k = 0; k < kernelSize; ++k)
-                    dst.convW[convOff + k * 9 + in * 3 + out] = take();
-
-        for (int out = 0; out < 3; ++out)
-            L.convB[out] = take();
-
-        for (int out = 0; out < 3; ++out)
-            L.mixinW[out] = take();
-
-        for (int out = 0; out < 3; ++out)
-            for (int in = 0; in < 3; ++in)
-                dst.l1x1W[l1Off + in * 3 + out] = take();
-
-        for (int out = 0; out < 3; ++out)
-            L.l1x1B[out] = take();
-
-        // First-layer current-tap precombine:
-        // z += mixin*x + W_current*(rechannel*x)
-        if (li == 0 && kernelSize == 6)
-        {
-            const float* wc = dst.convW + convOff + 5 * 9;
-            for (int out = 0; out < 3; ++out)
-                L.preCurrent[out] = L.mixinW[out]
-                                  + dst.rechannel[0] * wc[0 * 3 + out]
-                                  + dst.rechannel[1] * wc[1 * 3 + out]
-                                  + dst.rechannel[2] * wc[2 * 3 + out];
-        }
-
-        convOff += kernelSize * 9;
-        l1Off += 9;
-    }
-
-    for (int in = 0; in < 3; ++in)
-        for (int k = 0; k < kHeadKernel; ++k)
-            dst.headW[k][in] = take();
-
-    dst.headB = take();
-    dst.headScale = take();
-
-    dst.loaded = (p == count
-               && convOff == kConvWeightCount
-               && l1Off == kLayer1x1WeightCount
-               && kLayerHistoryOffset[kNumLayers - 1]
-                    + kLayerCols[kNumLayers - 1] * kChannels == kHistoryFloats);
-
-    return dst.loaded;
-}
-
-inline bool load_default_model(A2SharedWeights& dst) noexcept
-{
-    return load_weights(dst, kJcm2000A2NanoWeights, kA2WeightCount);
-}
-
-inline void reset_state(A2State& st, A2HotState& hot, const A2SharedWeights&) noexcept
-{
-    std::fill(std::begin(st.history), std::end(st.history), 0.0f);
-    std::fill(std::begin(hot.headHistory), std::end(hot.headHistory), 0.0f);
-    std::fill(std::begin(hot.bufA), std::end(hot.bufA), 0.0f);
-    std::fill(std::begin(hot.bufB), std::end(hot.bufB), 0.0f);
-    std::fill(std::begin(hot.headSum), std::end(hot.headSum), 0.0f);
-
-    for (int i = 0; i < kNumLayers; ++i)
-        st.layerWritePos[i] = kLayerCols[i] - 1;
-
-    hot.headWritePos = kHeadKernel - 1;
-}
-
-inline constexpr int prewarm_samples() noexcept
-{
-    int total = kHeadKernel - 1;
-    for (int i = 0; i < kNumLayers; ++i)
-        total += (kKernelSizes[i] - 1) * kDilations[i];
-    return total;
-}
-
-template<int K>
-NAM_A2_NOINLINE void process_layer_kernel(A2State& st,
-                                          A2HotState& hot,
-                                          const A2SharedWeights& sw,
-                                          int li,
-                                          const float* cond,
-                                          const float* in,
-                                          float* out) noexcept
-{
-    const A2LayerRuntime& L = sw.layer[li];
-
-    int wp = st.layerWritePos[li];
-    const int cols = kLayerCols[li];
-    const int dilation = kDilations[li];
-
-    float* const hist = st.history + kLayerHistoryOffset[li];
-    const float* const wAll = sw.convW + kLayerConvOffset[li];
-    const float* const lx = sw.l1x1W + kLayerL1Offset[li];
-
-    const float cb0 = L.convB[0];
-    const float cb1 = L.convB[1];
-    const float cb2 = L.convB[2];
-
-    const bool precombineCurrent = (li == 0 && K == 6);
-    const float mw0 = precombineCurrent ? L.preCurrent[0] : L.mixinW[0];
-    const float mw1 = precombineCurrent ? L.preCurrent[1] : L.mixinW[1];
-    const float mw2 = precombineCurrent ? L.preCurrent[2] : L.mixinW[2];
-
-    const float b0 = L.l1x1B[0];
-    const float b1 = L.l1x1B[1];
-    const float b2 = L.l1x1B[2];
-
-    for (int n = 0; n < kBlockSize; ++n)
-    {
-        const float* const srcIn = in + n * 3;
-        float* const histCol = hist + wp * 3;
-
-        const float x0 = srcIn[0];
-        const float x1 = srcIn[1];
-        const float x2 = srcIn[2];
-
-        histCol[0] = x0;
-        histCol[1] = x1;
-        histCol[2] = x2;
-
-        const float c = cond[n];
-        float z0 = cb0 + mw0 * c;
-        float z1 = cb1 + mw1 * c;
-        float z2 = cb2 + mw2 * c;
-
-        for (int k = 0; k < K; ++k)
-        {
-            if (precombineCurrent && k == K - 1)
-                continue;
-
-            const int lag = (K - 1 - k) * dilation;
-            int col = wp - lag;
-            if (col < 0)
-                col += cols;
-
-            const float* const s = hist + col * 3;
-            acc_tap3(wAll + k * 9, s, z0, z1, z2);
-        }
-
-        const float a0 = leaky(z0);
-        const float a1 = leaky(z1);
-        const float a2 = leaky(z2);
-
-        float* const hs = hot.headSum + n * 3;
-        hs[0] += a0;
-        hs[1] += a1;
-        hs[2] += a2;
-
-        float* const dst = out + n * 3;
-        dst[0] = x0 + b0 + a0 * lx[0] + a1 * lx[3] + a2 * lx[6];
-        dst[1] = x1 + b1 + a0 * lx[1] + a1 * lx[4] + a2 * lx[7];
-        dst[2] = x2 + b2 + a0 * lx[2] + a1 * lx[5] + a2 * lx[8];
-
-        ++wp;
-        if (wp >= cols)
-            wp = 0;
-    }
-
-    st.layerWritePos[li] = wp;
-}
-
-
-// Dual-sample K=6 kernel.
-// Requires kLayerCols = maxLookback + 2 so sample n+1 can be written before
-// sample n consumes its oldest tap. This improves FPU/load scheduling without
-// switching to large power-of-two rings.
-NAM_A2_NOINLINE void process_layer_kernel6_dual(A2State& st,
-                                                A2HotState& hot,
-                                                const A2SharedWeights& sw,
-                                                int li,
-                                                const float* __restrict__ cond,
-                                                const float* __restrict__ in,
-                                                float* __restrict__ out) noexcept
-{
-    const A2LayerRuntime& L = sw.layer[li];
-
-    int wp = st.layerWritePos[li];
-    const int cols = kLayerCols[li];
-    const int dilation = kDilations[li];
-
-    float* const hist = st.history + kLayerHistoryOffset[li];
-    const float* __restrict__ const wAll = sw.convW + kLayerConvOffset[li];
-    const float* __restrict__ const lx = sw.l1x1W + kLayerL1Offset[li];
-
-    const float cb0 = L.convB[0];
-    const float cb1 = L.convB[1];
-    const float cb2 = L.convB[2];
-
-    const bool precombineCurrent = (li == 0);
-    const float mw0 = precombineCurrent ? L.preCurrent[0] : L.mixinW[0];
-    const float mw1 = precombineCurrent ? L.preCurrent[1] : L.mixinW[1];
-    const float mw2 = precombineCurrent ? L.preCurrent[2] : L.mixinW[2];
-
-    const float b0 = L.l1x1B[0];
-    const float b1 = L.l1x1B[1];
-    const float b2 = L.l1x1B[2];
-
-    for (int n = 0; n < kBlockSize; n += 2)
-    {
-        const int wpA = wp;
-        int wpB = wpA + 1;
-        if (wpB >= cols)
-            wpB = 0;
-
-        const float* __restrict__ const srcA = in + n * 3;
-        const float* __restrict__ const srcB = srcA + 3;
-
-        const float ax0 = srcA[0];
-        const float ax1 = srcA[1];
-        const float ax2 = srcA[2];
-        const float bx0 = srcB[0];
-        const float bx1 = srcB[1];
-        const float bx2 = srcB[2];
-
-        float* const histA = hist + wpA * 3;
-        float* const histB = hist + wpB * 3;
-
-        histA[0] = ax0; histA[1] = ax1; histA[2] = ax2;
-        histB[0] = bx0; histB[1] = bx1; histB[2] = bx2;
-
-        const float ca = cond[n];
-        const float cb = cond[n + 1];
-
-        float za0 = cb0 + mw0 * ca;
-        float za1 = cb1 + mw1 * ca;
-        float za2 = cb2 + mw2 * ca;
-        float zb0 = cb0 + mw0 * cb;
-        float zb1 = cb1 + mw1 * cb;
-        float zb2 = cb2 + mw2 * cb;
-
-        // K=6 current tap optimization:
-        // For layers other than layer 0, tap 5 has lag 0, so it reads the
-        // current input we just wrote to history. Use srcA/srcB directly and
-        // loop over historical taps 0..4 only. Layer 0 already folds mixin +
-        // current tap into preCurrent, so it also uses only taps 0..4.
-        if (!precombineCurrent)
-        {
-            const float* __restrict__ const wc = wAll + 5 * 9;
-            za0 += ax0 * wc[0] + ax1 * wc[3] + ax2 * wc[6];
-            za1 += ax0 * wc[1] + ax1 * wc[4] + ax2 * wc[7];
-            za2 += ax0 * wc[2] + ax1 * wc[5] + ax2 * wc[8];
-
-            zb0 += bx0 * wc[0] + bx1 * wc[3] + bx2 * wc[6];
-            zb1 += bx0 * wc[1] + bx1 * wc[4] + bx2 * wc[7];
-            zb2 += bx0 * wc[2] + bx1 * wc[5] + bx2 * wc[8];
-        }
-
-        for (int k = 0; k < 5; ++k)
-        {
-            const int lag = (5 - k) * dilation;
-
-            int colA = wpA - lag;
-            if (colA < 0)
-                colA += cols;
-
-            int colB = wpB - lag;
-            if (colB < 0)
-                colB += cols;
-
-            const float* __restrict__ const sA = hist + colA * 3;
-            const float* __restrict__ const sB = hist + colB * 3;
-            const float* __restrict__ const w = wAll + k * 9;
-
-            acc_tap3(w, sA, za0, za1, za2);
-            acc_tap3(w, sB, zb0, zb1, zb2);
-        }
-
-        const float aa0 = leaky(za0);
-        const float aa1 = leaky(za1);
-        const float aa2 = leaky(za2);
-        const float ab0 = leaky(zb0);
-        const float ab1 = leaky(zb1);
-        const float ab2 = leaky(zb2);
-
-        float* __restrict__ const hsA = hot.headSum + n * 3;
-        float* __restrict__ const hsB = hsA + 3;
-
-        hsA[0] += aa0; hsA[1] += aa1; hsA[2] += aa2;
-        hsB[0] += ab0; hsB[1] += ab1; hsB[2] += ab2;
-
-        float* __restrict__ const dstA = out + n * 3;
-        float* __restrict__ const dstB = dstA + 3;
-
-        dstA[0] = ax0 + b0 + aa0 * lx[0] + aa1 * lx[3] + aa2 * lx[6];
-        dstA[1] = ax1 + b1 + aa0 * lx[1] + aa1 * lx[4] + aa2 * lx[7];
-        dstA[2] = ax2 + b2 + aa0 * lx[2] + aa1 * lx[5] + aa2 * lx[8];
-
-        dstB[0] = bx0 + b0 + ab0 * lx[0] + ab1 * lx[3] + ab2 * lx[6];
-        dstB[1] = bx1 + b1 + ab0 * lx[1] + ab1 * lx[4] + ab2 * lx[7];
-        dstB[2] = bx2 + b2 + ab0 * lx[2] + ab1 * lx[5] + ab2 * lx[8];
-
-        wp = wpB + 1;
-        if (wp >= cols)
-            wp = 0;
-    }
-
-    st.layerWritePos[li] = wp;
-}
-
-NAM_A2_NOINLINE void process_head(A2HotState& hot,
-                                  const A2SharedWeights& sw,
-                                  float* output) noexcept
-{
-    int wp = hot.headWritePos;
-    float* const hist = hot.headHistory;
-
-    for (int n = 0; n < kBlockSize; ++n)
-    {
-        const float* const hs = hot.headSum + n * 3;
-        float* const hcol = hist + wp * 3;
-
-        hcol[0] = hs[0];
-        hcol[1] = hs[1];
-        hcol[2] = hs[2];
-
-        float y = sw.headB;
-
-        for (int tap = 0; tap < kHeadKernel; ++tap)
-        {
-            const int lag = kHeadKernel - 1 - tap;
-            int col = wp - lag;
-            if (col < 0)
-                col += kHeadKernel;
-
-            const float* const s = hist + col * 3;
-            y += sw.headW[tap][0] * s[0]
-               + sw.headW[tap][1] * s[1]
-               + sw.headW[tap][2] * s[2];
-        }
-
-        output[n] = y * sw.headScale;
-
-        ++wp;
-        if (wp >= kHeadKernel)
-            wp = 0;
-    }
-
-    hot.headWritePos = wp;
-}
-
-NAM_A2_NOINLINE void process_block_48(A2State& st,
-                                      A2HotState& hot,
-                                      const A2SharedWeights& sw,
-                                      const float* input,
-                                      float* output) noexcept
-{
-    if (!sw.loaded || input == nullptr || output == nullptr)
-    {
-        if (output != nullptr)
-            std::fill(output, output + kBlockSize, 0.0f);
-        return;
-    }
-
-    float* in = hot.bufA;
-    float* out = hot.bufB;
-
-    const float r0 = sw.rechannel[0];
-    const float r1 = sw.rechannel[1];
-    const float r2 = sw.rechannel[2];
-
-    for (int n = 0; n < kBlockSize; ++n)
-    {
-        const float x = input[n];
-        float* const dst = in + n * 3;
-        dst[0] = r0 * x;
-        dst[1] = r1 * x;
-        dst[2] = r2 * x;
-
-        float* const hs = hot.headSum + n * 3;
-        hs[0] = 0.0f;
-        hs[1] = 0.0f;
-        hs[2] = 0.0f;
-    }
-
-    for (int li = 0; li < kNumLayers; ++li)
-    {
-        if (kKernelSizes[li] == 6)
-            process_layer_kernel6_dual(st, hot, sw, li, input, in, out);
-        else
-            process_layer_kernel<15>(st, hot, sw, li, input, in, out);
-
-        float* tmp = in;
-        in = out;
-        out = tmp;
-    }
-
-    process_head(hot, sw, output);
-}
-
-inline void prewarm(A2State& st, A2HotState& hot, const A2SharedWeights& sw) noexcept
-{
-    reset_state(st, hot, sw);
-    float zeros[kBlockSize] {};
-    float discard[kBlockSize] {};
-    int remaining = prewarm_samples();
-    while (remaining > 0)
-    {
-        process_block_48(st, hot, sw, zeros, discard);
-        remaining -= kBlockSize;
-    }
-}
-
-class A2Jcm2000Daisy48
-{
-public:
-    bool load_default_model() noexcept
-    {
-        const bool ok = nam_a2_daisy::load_default_model(weights_);
-        reset_state(state_, hot_, weights_);
-        prewarm(state_, hot_, weights_);
-        return ok;
-    }
-
-    bool load_weights(const float* weights, std::size_t count) noexcept
-    {
-        const bool ok = nam_a2_daisy::load_weights(weights_, weights, count);
-        reset_state(state_, hot_, weights_);
-        prewarm(state_, hot_, weights_);
-        return ok;
-    }
-
-    void reset() noexcept
-    {
-        reset_state(state_, hot_, weights_);
-        if(weights_.loaded)
-            prewarm(state_, hot_, weights_);
-    }
-
-    void process_block_48(const float* input, float* output) noexcept
-    {
-        nam_a2_daisy::process_block_48(state_, hot_, weights_, input, output);
-    }
-
-    A2SharedWeights& weights() noexcept { return weights_; }
-    A2State& state() noexcept { return state_; }
-    A2HotState& hot_state() noexcept { return hot_; }
-
-private:
-    // Shared packed model data is small enough for DTCM.
-    // Mutable history/work buffers remain per-instance in normal SRAM.
-    static inline A2SharedWeights weights_ ;
-    static inline A2HotState hot_ ;
-    A2State state_;
-};
-
-} // namespace nam_a2_daisy
